+++ src/bot.py
@@ -603,14 +603,14 @@ def main():
         None
     """
     # Thread for processing queue
-    thread_queue_handler = threading.Thread(target=queue_handler_thread, args=(), name="Thread-queue-handler")
+    thread_queue_handler = threading.Thread(target=queue_handler_thread, args=(), name="QueueHandlerThread")
     thread_queue_handler.start()
     # Thread for update status message
-    thread_status_message = threading.Thread(target=status_message_updater_thread, args=(), name="Thread-message-updater")
+    thread_status_message = threading.Thread(target=status_message_updater_thread, args=(), name="MessageUpdaterThread")
     thread_status_message.start()
     # Thread for export metrics
-    threads = [thread_queue_handler, thread_status_message]
-    thread_metrics = threading.Thread(target=metrics.run, args=(threads,), name="Thread-metrics")
+    threads = threading.enumerate()
+    thread_metrics = threading.Thread(target=metrics.run, args=(threads,), name="MetricsThread")
     thread_metrics.start()
     # Run bot
     while True: